Move and improve tvdb http request methods
Move tvdb specific request functions into a tvdb type struct that can
reuse the jwt token after calling the Login() function. Makes the
methods easier to use as you don't have to pass in the token each time.

Improved the request methods by returning a pointer to the specific type
of response that will be mapped. Previously, the responses were mapped
to a success and error interface that a user passed in. This was very
user error prone.