Sort and limit series list
Sort series list by it's priority which for tvdb is it's rating and site count. This implies that more popular shows would appear closer to the start of the list. Also limited the number of series shown at a time, in case the user's intended selection is high on the sorted list.